+++ packages/core-js/modules/esnext.async-iterator.flat-map.js
@@ -3,7 +3,6 @@
 var $ = require('../internals/export');
 var aFunction = require('../internals/a-function');
 var anObject = require('../internals/an-object');
-var isObject = require('../internals/is-object');
 var createAsyncIteratorProxy = require('../internals/create-async-iterator-proxy');
 var getAsyncIteratorMethod = require('../internals/get-async-iterator-method');
 
@@ -23,11 +22,12 @@ var AsyncIteratorProxy = createAsyncIteratorProxy(function (arg, Promise) {
             } else {
               Promise.resolve(mapper(step.value)).then(function (mapped) {
                 try {
-                  if (isObject(mapped) && (iteratorMethod = getAsyncIteratorMethod(mapped)) !== undefined) {
-                    state.innerIterator = innerIterator = iteratorMethod.call(mapped);
+                  iteratorMethod = getAsyncIteratorMethod(mapped);
+                  if (iteratorMethod !== undefined) {
+                    state.innerIterator = innerIterator = anObject(iteratorMethod.call(mapped));
                     state.innerNext = aFunction(innerIterator.next);
                     return innerLoop();
-                  } resolve({ done: false, value: mapped });
+                  } reject(TypeError('.flatMap callback should return an iterable object'));
                 } catch (error2) { reject(error2); }
               }, reject);
             }