+++ tests/tests/esnext.async-iterator.flat-map.js
@@ -12,12 +12,13 @@ QUnit.test('AsyncIterator#flatMap', assert => {
   assert.looksNative(flatMap);
   assert.nonEnumerable(AsyncIterator.prototype, 'flatMap');
 
-  flatMap.call(createIterator([1, [], 2, createIterable([3, 4]), [5, 6], 'ab']), it => typeof it == 'number' ? -it : it).toArray().then(it => {
-    assert.arrayEqual(it, [-1, -2, 3, 4, 5, 6, 'ab'], 'basic functionality');
+  flatMap.call(createIterator([1, [], 2, createIterable([3, 4]), [5, 6], 'ab']), it => typeof it == 'number' ? [-it] : it).toArray().then(it => {
+    assert.arrayEqual(it, [-1, -2, 3, 4, 5, 6, 'a', 'b'], 'basic functionality');
     return flatMap.call(createIterator([1]), function (arg) {
       assert.same(this, STRICT_THIS, 'this');
       assert.same(arguments.length, 1, 'arguments length');
       assert.same(arg, 1, 'argument');
+      return [arg];
     }).toArray();
   }).then(() => {
     return flatMap.call(createIterator([1]), () => { throw 42; }).toArray();